# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LeadStatus"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead with events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadWithEvents"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadInput"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/events:
    post:
      operationId: createLeadEvent
      tags: [leads]
      summary: Add an event/note to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadEvent"
...
  /leads/{id}/score:
    get:
      operationId: getLeadScore
      tags: [leads]
      summary: Get predictive score for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead score and factor breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadScore"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session JWT. The browser sends this as a cookie managed by the Clerk SDK; mobile clients send it in the Authorization header as `Bearer <token>`.
  schemas:
...
    LeadStatus:
      type: string
      enum: [new, contacted, interested, won, lost]
    ScoreFactors:
      type: object
      properties:
        visitRecency:
          type: number
        status:
          type: number
        activity:
          type: number
        outreachEngagement:
          type: number
        sequenceProgress:
          type: number
      required:
        - visitRecency
        - status
        - activity
        - outreachEngagement
        - sequenceProgress
    LeadScore:
      type: object
      properties:
        score:
          type: integer
        factors:
          $ref: "#/components/schemas/ScoreFactors"
      required:
        - score
        - factors
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        visitDate:
          type: string
        status:
          $ref: "#/components/schemas/LeadStatus"
        notes:
          type: string
          nullable: true
        score:
          type: integer
        scoreFactors:
          $ref: "#/components/schemas/ScoreFactors"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - visitDate
        - status
        - score
        - createdAt
        - updatedAt
    LeadEvent:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        type:
          type: string
        note:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - leadId
        - type
        - createdAt
    LeadWithEvents:
      allOf:
        - $ref: "#/components/schemas/Lead"
        - type: object
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/LeadEvent"
          required:
            - events
    CreateLeadInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        visitDate:
          type: string
        status:
          $ref: "#/components/schemas/LeadStatus"
        notes:
          type: string
          nullable: true
      required:
        - name
        - email
        - phone
        - visitDate
    UpdateLeadInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        visitDate:
          type: string
        status:
          $ref: "#/components/schemas/LeadStatus"
        notes:
          type: string
          nullable: true
    CreateLeadEventInput:
      type: object
      properties:
        type:
          type: string
        note:
          type: string
          nullable: true
      required:
        - type
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...